Start and stop jobs from the model
