+++ bookwyrm/importers/importer.py
@@ -1,15 +1,7 @@
 """ handle reading a csv from an external service, defaults are from Goodreads """
 import csv
-import logging
-
 from django.utils import timezone
-from django.utils.translation import gettext_lazy as _
-
-from bookwyrm import models
 from bookwyrm.models import ImportJob, ImportItem
-from bookwyrm.tasks import app, LOW
-
-logger = logging.getLogger(__name__)
 
 
 class Importer:
@@ -118,135 +110,3 @@ class Importer:
             # this will re-normalize the raw data
             self.create_item(job, item.index, item.data)
         return job
-
-    def start_import(self, job):  # pylint: disable=no-self-use
-        """initalizes a csv import job"""
-        result = start_import_task.delay(job.id)
-        job.task_id = result.id
-        job.save()
-
-
-@app.task(queue="low_priority")
-def start_import_task(job_id):
-    """trigger the child tasks for each row"""
-    job = ImportJob.objects.get(id=job_id)
-    # these are sub-tasks so that one big task doesn't use up all the memory in celery
-    for item in job.items.values_list("id", flat=True).all():
-        task = import_item_task.delay(item)
-        item.task_id = task.id
-        item.save()
-    job.status = "active"
-    job.save()
-
-
-@app.task(queue="low_priority")
-def import_item_task(item_id):
-    """resolve a row into a book"""
-    item = models.ImportItem.objects.get(id=item_id)
-    # make sure the job has not been stopped
-    if item.job.complete:
-        return
-
-    try:
-        item.resolve()
-    except Exception as err:  # pylint: disable=broad-except
-        item.fail_reason = _("Error loading book")
-        item.save()
-        item.update_job()
-        raise err
-
-    if item.book:
-        # shelves book and handles reviews
-        handle_imported_book(item)
-    else:
-        item.fail_reason = _("Could not find a match for book")
-
-    item.save()
-    item.update_job()
-
-
-def handle_imported_book(item):
-    """process a csv and then post about it"""
-    job = item.job
-    user = job.user
-    if isinstance(item.book, models.Work):
-        item.book = item.book.default_edition
-    if not item.book:
-        item.fail_reason = _("Error loading book")
-        item.save()
-        return
-    if not isinstance(item.book, models.Edition):
-        item.book = item.book.edition
-
-    existing_shelf = models.ShelfBook.objects.filter(book=item.book, user=user).exists()
-
-    # shelve the book if it hasn't been shelved already
-    if item.shelf and not existing_shelf:
-        desired_shelf = models.Shelf.objects.get(identifier=item.shelf, user=user)
-        shelved_date = item.date_added or timezone.now()
-        models.ShelfBook(
-            book=item.book, shelf=desired_shelf, user=user, shelved_date=shelved_date
-        ).save(priority=LOW)
-
-    for read in item.reads:
-        # check for an existing readthrough with the same dates
-        if models.ReadThrough.objects.filter(
-            user=user,
-            book=item.book,
-            start_date=read.start_date,
-            finish_date=read.finish_date,
-        ).exists():
-            continue
-        read.book = item.book
-        read.user = user
-        read.save()
-
-    if job.include_reviews and (item.rating or item.review) and not item.linked_review:
-        # we don't know the publication date of the review,
-        # but "now" is a bad guess
-        published_date_guess = item.date_read or item.date_added
-        if item.review:
-            # pylint: disable=consider-using-f-string
-            review_title = "Review of {!r} on {!r}".format(
-                item.book.title,
-                job.source,
-            )
-            review = models.Review.objects.filter(
-                user=user,
-                book=item.book,
-                name=review_title,
-                rating=item.rating,
-                published_date=published_date_guess,
-            ).first()
-            if not review:
-                review = models.Review(
-                    user=user,
-                    book=item.book,
-                    name=review_title,
-                    content=item.review,
-                    rating=item.rating,
-                    published_date=published_date_guess,
-                    privacy=job.privacy,
-                )
-                review.save(software="bookwyrm", priority=LOW)
-        else:
-            # just a rating
-            review = models.ReviewRating.objects.filter(
-                user=user,
-                book=item.book,
-                published_date=published_date_guess,
-                rating=item.rating,
-            ).first()
-            if not review:
-                review = models.ReviewRating(
-                    user=user,
-                    book=item.book,
-                    rating=item.rating,
-                    published_date=published_date_guess,
-                    privacy=job.privacy,
-                )
-                review.save(software="bookwyrm", priority=LOW)
-
-        # only broadcast this review to other bookwyrm instances
-        item.linked_review = review
-    item.save()
